Refactor compression fields from double to float in aggregate and schema structures
This commit updates the compression fields in various structures, changing their type from double to float for consistency and improved memory usage. Affected files include aggrexpr.h, columninfo.h, sphinx.h, and sphinxjsonquery.cpp, ensuring all related functions and methods are adjusted accordingly.

+++ src/aggrexpr.h
@@ -92,19 +92,19 @@ struct AggrPercentilesSetting_t
 {
 	CSphVector<float> m_dPercents;
 	bool m_bKeyed = false;
-	double m_fCompression = 200.0;
+	float m_fCompression = 200.0f;
 };
 
 struct AggrPercentileRanksSetting_t
 {
 	CSphVector<double> m_dValues;
 	bool m_bKeyed = false;
-	double m_fCompression = 200.0;
+	float m_fCompression = 200.0f;
 };
 
 struct AggrMadSetting_t
 {
-	double m_fCompression = 1000.0;
+	float m_fCompression = 1000.0f;
 };
 
 struct AggrSettings_t
